+++ src/main/java/org/riekr/jloga/io/VolatileTextSource.java
@@ -1,29 +1,78 @@
 package org.riekr.jloga.io;
 
-import java.util.Arrays;
+import java.util.Collections;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
+import java.util.function.IntConsumer;
+
+import org.riekr.jloga.react.IntBehaviourSubject;
+import org.riekr.jloga.react.Observer;
+import org.riekr.jloga.react.Unsubscribable;
 
 public class VolatileTextSource implements TextSource {
 
-	protected final List<String> _data;
+	private final IntBehaviourSubject _lineCountSubject = new IntBehaviourSubject();
+	private final List<String>        _data             = new LinkedList<>();
+	private final ReadWriteLock       _readWriteLock    = new ReentrantReadWriteLock();
 
-	public VolatileTextSource(String... data) {
-		this(Arrays.asList(data));
+	public VolatileTextSource() {}
+
+	public VolatileTextSource(String... lines) {
+		Collections.addAll(_data, lines);
 	}
 
-	public VolatileTextSource(List<String> data) {
-		_data = data;
+	public final void clear() {
+		final Lock lock = _readWriteLock.writeLock();
+		lock.lock();
+		try {
+			_data.clear();
+		} finally {
+			lock.unlock();
+		}
+		_lineCountSubject.next(0);
+	}
+
+	public final void add(String line) {
+		final Lock lock = _readWriteLock.writeLock();
+		lock.lock();
+		try {
+			_data.add(line);
+		} finally {
+			lock.unlock();
+		}
+		_lineCountSubject.next(_data.size());
 	}
 
 	@Override
 	public String getText(int line) {
-		if (line < 0 || line >= _data.size())
-			return "";
-		return _data.get(line);
+		final Lock lock = _readWriteLock.readLock();
+		lock.lock();
+		try {
+			if (line < 0 || line >= _data.size())
+				return "";
+			return _data.get(line);
+		} finally {
+			lock.unlock();
+		}
 	}
 
 	@Override
-	public int getLineCount() {
+	public Future<Integer> requestIntermediateLineCount(IntConsumer consumer) {
+		return _lineCountSubject.once(Observer.async(consumer::accept));
+	}
+
+	@Override
+	public Unsubscribable subscribeLineCount(IntConsumer consumer) {
+		return _lineCountSubject.subscribe(consumer::accept);
+	}
+
+	@Override
+	public int getLineCount() throws ExecutionException, InterruptedException {
 		return _data.size();
 	}
 }